Travis: Fixed ESLint syntax tests not working for ES6 files
This has been fixed by delaying the installation of the style config files until the style tests are run and re-enabling the usage of .eslintrc* files for ESLint syntax tests. This way, the .eslintrc* files that declare the JavaScript environments for various subdirectories can be used to inform which syntax tests are run.