Update composer check to be more granular
Currently if any portion of `composer.json` is modified then `composer.lock`
must be modified or the process fails. Unfortunately, there are some times
when we would legitimately expect there to be no change in `composer.lock` when
`composer.json` is updated ( say, for instance, new entries are added to the
`classmap` section ).

It turns out that there are only a few areas of `composer.json` that we should
be inspecting to determine whether or not `composer.lock` should also be
updated.

This commit adds a new script `composer_check.py` that, when run, will compare
these sections for the current and original `composer.json`. If a difference is
found then its exit code will not be 0, else it will be 0.

`build.sh` has been updated to run this script and take the appropriate action
if there are changes detected.